requirements.txt
openai>=1.15.0
langchain>=0.1.16
langgraph>=0.0.26
langchain-core>=0.1.42
python-dotenv>=1.0.0
streamlit>=1.30.0
xlsxwriter>=3.1.9
PyPDF2>=3.0.1
pandas>=2.1.0
openpyxl>=3.1.2
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
pydantic>=2.5.0
